[UPDATE] Rearrange files
