+++ package.json
@@ -23,7 +23,7 @@
 		"helmet": "^7.1.0",
 		"hpp": "^0.2.3",
 		"jsonwebtoken": "^9.0.2",
-		"mongodb": "^6.3.0",
+		"mongodb": "^2.2.12",
 		"mongodb-memory-server": "^9.1.1",
 		"mongoose": "^8.0.2",
 		"rtcmulticonnection-server": "^1.3.3",
